refactor(fiber): move file to main
